ci: run tests on master as well
Previously we were only running on pull requests, but this means that we
don't get test results for the master branch. This is a problem because
we want to know if the master branch is broken before we perform a
release of the App. We could enforce that all PRs must be up to date
with master, but this is a bit of a pain and doesn't really solve the
problem of knowing for sure that master is broken.