Prepare tests for decide v4 work
Decide v4 changes the format for how flags are returned by the `/decide` API endpoint.

The existing tests and test files have been renamed to include v3 in the name because the library will need to be backwards compatible.

I also made the test responses more closely adhere to real world `/decide` responses. In particular:

1. In `featureFlagPayloads`, there were keys that were not in `featureFlags`. But in actuality, we only have payloads for active flags, so the key _must_ be in both places if there's a payload.
2. Feature flag payloads are always string encoded JSON or `null`.

I changed the files and tests to account for that.

Also removed an unused test file.